askely_whatsapp_render/requirements.txt
Flask==3.1.1
gunicorn==21.2.0
twilio==9.6.3
psycopg2-binary==2.9.9